refactor: add missing ng prefix to the angular module name
replace module name from 'translate-async-loader' to 'ngTranslateAsyncLoader'

BREAKING CHANGE: module is now named as `ngTranslateAsyncLoader`
