+++ README.md
@@ -1,24 +1,24 @@
 # Translate async loader
 
-> Angular translate asynchronous loader
+> AngularJS translate asynchronous loader
 
-[![Downloads](https://badgen.net/npm/dt/@ovh-ux/translate-async-loader)](https://npmjs.com/package/@ovh-ux/translate-async-loader) [![Dependencies](https://badgen.net/david/dep/ovh-ux/translate-async-loader)](https://npmjs.com/package/@ovh-ux/translate-async-loader?activeTab=dependencies) [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/translate-async-loader)](https://npmjs.com/package/@ovh-ux/translate-async-loader?activeTab=dependencies) [![Gitter](https://badgen.net/badge/gitter/ovh-ux/blue?icon=gitter)](https://gitter.im/ovh/ux)
+[![Downloads](https://badgen.net/npm/dt/@ovh-ux/ng-translate-async-loader)](https://npmjs.com/package/@ovh-ux/ng-translate-async-loader) [![Dependencies](https://badgen.net/david/dep/ovh-ux/ng-translate-async-loader)](https://npmjs.com/package/@ovh-ux/ng-translate-async-loader?activeTab=dependencies) [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/ng-translate-async-loader)](https://npmjs.com/package/@ovh-ux/ng-translate-async-loader?activeTab=dependencies) [![Gitter](https://badgen.net/badge/gitter/ovh-ux/blue?icon=gitter)](https://gitter.im/ovh/ux)
 
 ## Install
 
 ```sh
-yarn add @ovh-ux/translate-async-loader
+yarn add @ovh-ux/ng-translate-async-loader
 ```
 
 ## Usage
 
 ```js
 import angular from 'angular';
-import translateAsyncLoader from '@ovh-ux/translate-async-loader';
+import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
 
 angular
   .module('myApp', [
-    translateAsyncLoader,
+    ngTranslateAsyncLoader,
   ]);
 ```
 
@@ -30,7 +30,7 @@ yarn test
 
 ## Contributing
 
-Always feel free to help out! Whether it's [filing bugs and feature requests](https://github.com/ovh-ux/translate-async-loader/issues/new) or working on some of the [open issues](https://github.com/ovh-ux/translate-async-loader/issues), our [contributing guide](CONTRIBUTING.md) will help get you started.
+Always feel free to help out! Whether it's [filing bugs and feature requests](https://github.com/ovh-ux/ng-translate-async-loader/issues/new) or working on some of the [open issues](https://github.com/ovh-ux/ng-translate-async-loader/issues), our [contributing guide](CONTRIBUTING.md) will help get you started.
 
 ## License
