+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ovh-ux/translate-async-loader",
+  "name": "@ovh-ux/ng-translate-async-loader",
   "version": "1.0.8",
   "description": "Angular translate asynchronous loader",
   "keywords": [
@@ -8,14 +8,14 @@
     "ovh",
     "translate"
   ],
-  "repository": "ovh-ux/translate-async-loader",
+  "repository": "ovh-ux/ng-translate-async-loader",
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "files": [
     "dist"
   ],
   "main": "./dist/cjs/index.js",
-  "browser": "./dist/umd/translate-async-loader.js",
+  "browser": "./dist/umd/ng-translate-async-loader.js",
   "scripts": {
     "build": "rollup -c",
     "lint": "run-p lint:js lint:md",
